Moving stuff around, improving imports formatting
