+++ tests/integration/test_project_navigator.py
@@ -5,7 +5,7 @@ import logging
 import mock
 
 from group_project_v2.project_navigator import ViewTypes
-from group_project_v2.stage_components import StageState
+from group_project_v2.stage import StageState
 from tests.integration.base_test import SingleScenarioTestSuite
 from tests.integration.page_elements import NavigationViewElement, ResourcesViewElement, SubmissionsViewElement
 from tests.utils import KNOWN_USERS
